m002

+++ SwordAttack.java
@@ -1,7 +1,7 @@
 package HomeWork_ext_RoleGame;
 
 public class SwordAttack implements IAttackBehavior {
-	public float num;
+	public double num;
 	public void attack() {
 		System.out.println("攻擊");
 	}
